Create a prototype object for primitives.
