+++ package.json
@@ -48,7 +48,7 @@
   },
   "dependencies": {
     "css-element-queries": "^1.2.2",
-    "lodash": "^4.17.15",
+    "lodash": "^4.17.19",
     "promise-polyfill": "^8.1.3",
     "three": "^0.105.0",
     "url-loader": "^2.1.0",
