+++ src/geometry.js
@@ -6,24 +6,6 @@ function absNumericalSort( a, b ) {
 
 }
 
-function PhongToToon(materialIn) {
-	if (materialIn.isMeshPhongMaterial) {
-		let material = new THREE.MeshToonMaterial({
-			color : materialIn.color.clone(),
-			morphTargets : materialIn.morphTargets,
-			morphNormals : materialIn.morphNormals,
-			vertexColors : materialIn.vertexColors,
-			transparent : materialIn.transparent,
-			opacity : materialIn.opacity,
-			side : materialIn.side
-		});
-		if (materialIn.map)
-			material.map = materialIn.map;
-		return material;
-	}
-
-	return materialIn;
-}
 
 /**
  * Provides an object which stores geometry and provides method which controls its animations.
@@ -34,32 +16,13 @@ function PhongToToon(materialIn) {
  * @author Alan Wu
  * @return {Zinc.Geometry}
  */
-exports.Geometry = function () {
+const Geometry = function () {
+  (require('./primitives/zincObject').ZincObject).call(this);
 	// THREE.Geometry or THREE.BufferGeometry
-	this.geometry = undefined;
-	this.mixer = undefined;
-	this.timeEnabled = false;
-	this.morphColour = false;
 	this.modelId = -1;
 	this.videoHandler = undefined;
-	// THREE.Mesh
-	this.morph = undefined;
-	this.clipAction = undefined;
   this.isGeometry = true;
-  this.marker = undefined;
-  let markerUpdateRequired = true;
-	/**
-	 * Total duration of the animation, this value interacts with the 
-	 * {@link Zinc.Renderer#playRate} to produce the actual duration of the
-	 * animation. Actual time in second = duration / playRate.
-	 */
-	this.duration = 3000;
-	/**
-	 * Groupname given to this geometry.
-	 */
-	this.groupName = undefined;
-	let inbuildTime = 0;
-	this.userData = [];
+
 
 	this.setMesh = (mesh, modelId, localTimeEnabled, localMorphColour) => {
 		this.mixer = new THREE.AnimationMixer(mesh);
@@ -155,105 +118,6 @@ exports.Geometry = function () {
 		this.setMesh(mesh, options.modelId, options.localTimeEnabled,
 			options.localMorphColour);
 	}
-
-	this.setName = groupNameIn => {
-		this.groupName = groupNameIn;
-		if (this.morph) {
-			this.morph.name = this.groupName;
-		}
-  }
-  
-  /**
-	 * Get the visibility of this Geometry.
-	 * 
-	 */
-	this.getVisibility = () => {
-		return this.morph.visible;
-	}
-	
-	/**
-	 * Set the visibility of this Geometry.
-	 * 
-	 * @param {Boolean} visible - a boolean flag indicate the visibility to be set 
-	 */
-	this.setVisibility = visible => {
-		this.morph.visible = visible;
-	}
-	
-	/**
-	 * Set the opacity of this Geometry. This function will also set the isTransparent
-	 * flag according to the provided alpha value.
-	 * 
-	 * @param {Number} alpah - Alpha value to set for this geometry, 
-	 * can be any value between from 0 to 1.0.
-	 */
-	this.setAlpha = alpha => {
-		const material = this.morph.material;
-		let isTransparent = false;
-		if (alpha  < 1.0)
-			isTransparent = true;
-		material.transparent = isTransparent;
-		material.opacity = alpha;
-	}
-
-	this.setFrustumCulled = flag => {
-		if (this.morph)
-			this.morph.frustumCulled = flag;
-	}
-	
-	/**
-	 * Get the local time of this geometry, it returns a value between 
-	 * 0 and the duration.
-	 * 
-	 * @return {Number}
-	 */
-	this.getCurrentTime = () => {
-		if (this.clipAction) {
-			const ratio = this.clipAction.time / this.clipAction._clip.duration;
-			return this.duration * ratio;
-		} else {
-			return inbuildTime;
-		}
-	}
-	
-	/**
-	 * Set the local time of this geometry.
-	 * 
-	 * @param {Number} time - Can be any value between 0 to duration.
-	 */
-	this.setMorphTime = time => {
-		let timeChanged = false;
-		if (this.clipAction) {
-			const ratio = time / this.duration;
-			const actualDuration = this.clipAction._clip.duration;
-			let newTime = ratio * actualDuration;
-			if (newTime != this.clipAction.time) {
-				this.clipAction.time = newTime;
-				timeChanged = true;
-			}
-			if (this.clipAction.time > actualDuration)
-				this.clipAction.time = actualDuration;
-			if (this.clipAction.time < 0.0)
-				this.clipAction.time = 0.0;
-			if (timeChanged && this.timeEnabled == 1)
-				this.mixer.update( 0.0 );
-		} else {
-			let newTime = time; 
-			if (time > this.duration)
-				newTime = this.duration;
-			else if (0 > time)
-				newTime = 0;
-			else
-				newTime = time;
-			if (newTime != inbuildTime) {
-				inbuildTime = newTime;
-				timeChanged = true;
-			}
-		}
-		if (timeChanged) {
-			updateMorphColorAttribute(this.geometry, this.morph, this.clipAction);
-		}
-	}
 	
 	this.calculateUVs = () => {
 		this.geometry.computeBoundingBox();
@@ -284,226 +148,7 @@ exports.Geometry = function () {
 		this.morph.material.wireframe = wireframe;
 	}
 	
-	this.setVertexColors = vertexColors => {
-		this.morph.material.vertexColors = vertexColors;
-		this.geometry.colorsNeedUpdate = true;
-	}
-	
-	/**
-	 * Set the colour of the geometry.
-	 * 
-	 * @param {THREE.Color} colour - Colour to be set for this geometry.
-	 */
-	this.setColour = colour => {
-		this.morph.material.color = colour;
-		this.geometry.colorsNeedUpdate = true;
-	}
-
-	this.getColourHex = () => {
-		if (!this.morphColour)
-			return this.morph.material.color.getHexString();
-		return undefined;
-	}
-
-	this.setColourHex = hex => {
-		this.morph.material.color.setHex(hex);
-	}
-	
-	/**
-	 * Set the material of the geometry.
-	 * 
-	 * @param {THREE.Material} material - Material to be set for this geometry.
-	 */
-	this.setMaterial = material => {
-		this.morph.material = material;
-		this.geometry.colorsNeedUpdate = true;
-	}
-
-	const updateMorphColorAttribute = (targetGeometry, morph, clipAction) => {
-		if (morph && targetGeometry.morphAttributes[ "color" ]) {
-			const morphColors = targetGeometry.morphAttributes[ "color" ];
-			const influences = morph.morphTargetInfluences;
-			const length = influences.length;
-			targetGeometry.removeAttribute( 'morphColor0' );
-			targetGeometry.removeAttribute( 'morphColor1' );
-			let bound = 0;
-			let morphArray = [];
-			for (let i = 0; (1 > bound) || (i < length); i++) {
-				if (influences[i] > 0) {
-					bound++;
-					morphArray.push([i, influences[i]]);
-				}
-			}
-			morphArray.sort(absNumericalSort);
-			if (morphArray.length == 2) {
-				targetGeometry.addAttribute('morphColor0', morphColors[ morphArray[0][0] ] );
-				targetGeometry.addAttribute('morphColor1', morphColors[ morphArray[1][0] ] );
-			} else if (morphArray.length == 1) {
-				targetGeometry.addAttribute('morphColor0', morphColors[ morphArray[0][0] ] );
-				targetGeometry.addAttribute('morphColor1', morphColors[ morphArray[0][0] ] );
-			}
-		}
-	}
-		
-	//Calculate the interpolated colour at current time
-	const morphColorsToVertexColors = (targetGeometry, morph, clipAction) => {
-		if ( morph && targetGeometry.morphColors && targetGeometry.morphColors.length) {
-			const getColorsRGB = require("./utilities").getColorsRGB;
-
-			let current_time = 0.0;
-			if (clipAction)
-				current_time = clipAction.time/clipAction._clip.duration * (targetGeometry.morphColors.length - 1);
-			else
-				current_time = inbuildTime/this.duration * (targetGeometry.morphColors.length - 1);
-			
-			const bottom_frame =  Math.floor(current_time);
-			const proportion = 1 - (current_time - bottom_frame);
-			const top_frame =  Math.ceil(current_time);
-			const bottomColorMap = targetGeometry.morphColors[ bottom_frame ];
-			const TopColorMap = targetGeometry.morphColors[ top_frame ];
-			
-			for ( let i = 0; i < targetGeometry.faces.length; i ++ ) {
-				let my_color1 = getColorsRGB(bottomColorMap.colors, targetGeometry.faces[i].a);
-				let my_color2 = getColorsRGB(TopColorMap.colors, targetGeometry.faces[i].a);
-				let resulting_color = [my_color1[0] * proportion + my_color2[0] * (1 - proportion),
-					my_color1[1] * proportion + my_color2[1] * (1 - proportion),
-					my_color1[2] * proportion + my_color2[2] * (1 - proportion)];
-				targetGeometry.faces[i].vertexColors[0].setRGB(resulting_color[0], resulting_color[1], resulting_color[2])
-				my_color1 = getColorsRGB(bottomColorMap.colors, targetGeometry.faces[i].b);
-				my_color2 = getColorsRGB(TopColorMap.colors, targetGeometry.faces[i].b);
-				resulting_color = [my_color1[0] * proportion + my_color2[0] * (1 - proportion),
-					my_color1[1] * proportion + my_color2[1] * (1 - proportion),
-					my_color1[2] * proportion + my_color2[2] * (1 - proportion)]
-				targetGeometry.faces[i].vertexColors[1].setRGB(resulting_color[0], resulting_color[1], resulting_color[2])
-				my_color1 = getColorsRGB(bottomColorMap.colors, targetGeometry.faces[i].c);
-				my_color2 = getColorsRGB(TopColorMap.colors, targetGeometry.faces[i].c);
-				resulting_color = [my_color1[0] * proportion + my_color2[0] * (1 - proportion),
-					my_color1[1] * proportion + my_color2[1] * (1 - proportion),
-					my_color1[2] * proportion + my_color2[2] * (1 - proportion)]
-				targetGeometry.faces[i].vertexColors[2].setRGB(resulting_color[0], resulting_color[1], resulting_color[2])
-			}	
-		}
-	}
-	
-  /**
-   * Check if the geometry is time varying.
-   * 
-   * @return {Boolean}
-   */
-	this.isTimeVarying = () => {
-		if (this.timeEnabled || this.morphColour)
-			return true;
-		return false;
-	}
-  
-	/**
-	 * Get the bounding box of this geometry.
-	 * 
-	 * @return {THREE.Box3}.
-	 */
-	this.getBoundingBox = () => {
-		if (this.morph && this.morph.visible) {
-			//var boundingBox = new THREE.Box3().setFromObject(this.morph);
-      //return boundingBox;
-      let influences = this.morph.morphTargetInfluences;
-      let attributes = this.morph.geometry.morphAttributes;
-      if (influences && attributes && attributes.position) {
-        let min = new THREE.Vector3();
-        let max = new THREE.Vector3();
-        let found = false;
-        for (let i = 0; i < influences.length; i++) {
-          if (influences[i] > 0) {
-            found = true;
-            let box = new THREE.Box3().setFromArray(
-              attributes.position[i].array);
-            min.add(box.min.multiplyScalar(influences[i]));
-            max.add(box.max.multiplyScalar(influences[i]));
-          }
-        }
-        if (found) {
-          let boundingBox = new THREE.Box3(min, max);
-          return boundingBox;
-        }
-      }
-      var boundingBox = new THREE.Box3().setFromObject(this.morph);
-      return boundingBox;
-		}
-		return undefined;
-	}
-			
-	/**
-	 * Clear this geometry and free the memory.
-	 */
-	this.dispose = () => {
-	  if (this.morph && this.morph.geometry)
-	    this.morph.geometry.dispose();
-	  if (this.morph && this.morph.material)
-	    this.morph.material.dispose();
-		this.geometry = undefined;
-		this.mixer = undefined;
-		this.morph = undefined;
-		this.clipAction = undefined;
-		this.groupName = undefined;
-  }
-  
-  let updateMarker = (playAnimation, options) => {
-    if (playAnimation == true || 
-      (false == (options && options.displayMarkers)))
-		{
-      if (this.marker) {
-        this.marker.disable();
-      }
-      markerUpdateRequired = true;  
-    } else {
-      if (this.groupName) {
-        if (!this.marker) {
-          this.marker = new (require("./marker").Marker)(this);
-          markerUpdateRequired = true;
-        }
-        if (markerUpdateRequired) {
-          markerUpdateRequired = false;
-          this.marker.enable();
-          //Remove the marker to get the accurate box
-          this.morph.remove(this.marker.graphicsObject);
-          let center = new THREE.Vector3();
-          this.getBoundingBox().getCenter(center);
-          this.marker.setPosition(center.x, center.y, center.z);
-          this.morph.add(this.marker.graphicsObject);
-        }
-      }
-    }
-    if (this.marker && this.marker.isEnabled() && 
-      options && options.camera && options.displayMarkers) {
-      this.marker.updateDistanceBasedOpacity(options.camera.cameraObject);
-    }
-  }
-	
-	//Update the geometry and colours depending on the morph.
-	this.render = (delta, playAnimation, options) => {
-		if (playAnimation == true) 
-		{
-			if ((this.clipAction) && (this.timeEnabled == 1)) {
-				this.mixer.update( delta );
-			}
-			else {
-				let targetTime = inbuildTime + delta;
-				if (targetTime > this.duration)
-					targetTime = targetTime - this.duration;
-				inbuildTime = targetTime;
-			}
-			if (delta != 0 && this.morphColour == 1) {
-				if (typeof this.geometry !== "undefined") {
-					
-					if (this.morph.material.vertexColors == THREE.VertexColors)
-					{
-						let clipAction = undefined;
-						if (this.clipAction && (this.timeEnabled == 1))
-							clipAction = this.clipAction;
-						updateMorphColorAttribute(this.geometry, this.morph, clipAction);
-					}	
-				}
-      }
-    }
-    updateMarker(playAnimation, options);
-	}
 }
+
+Geometry.prototype = Object.create((require('./primitives/zincObject').ZincObject).prototype);
+exports.Geometry = Geometry;
