+++ src/lines.js
@@ -8,47 +8,9 @@ const THREE = require('three');
  * @author Alan Wu
  * @return {Zinc.Lines}
  */
-exports.Lines = function () {
-	this.morph = undefined;
-	this.groupName = undefined;
-	this.timeEnabled = false;
-	this.morphColour = false;
+const Lines = function () {
+  (require('./primitives/zincObject').ZincObject).call(this);
 	this.isLines = true;
-	let inbuildTime = 0;
-	/**
-	 * Total duration of the animation, this value interacts with the 
-	 * {@link Zinc.Renderer#playRate} to produce the actual duration of the
-	 * animation. Actual time in second = duration / playRate.
-	 */
-	this.duration = 3000;
-	this.clipAction = undefined;
-	this.userData = [];
-	
-	this.setFrustumCulled = flag => {
-		if (this.morph)
-			this.morph.frustumCulled = flag;
-	}
-		
-	/**
-	 * Check if the pointset is time varying.
-	 * 
-	 * @return {Boolean}
-	 */
-	this.isTimeVarying = () => {
-	if (this.timeEnabled || this.morphColour)
-		return true;
-	return false;
-	}
-
-	this.setColourHex = hex => {
-		this.morph.material.color.setHex(hex);
-	}
-
-	this.getColourHex = () => {
-		if (!this.morphColour)
-			return this.morph.material.color.getHexString();
-		return undefined;
-	}
 
 	this.setLine = (line, localTimeEnabled, localMorphColour) => {
 		this.mixer = new THREE.AnimationMixer(line);
@@ -96,168 +58,7 @@ exports.Lines = function () {
 		}
 	}
 
-	this.setName = groupNameIn => {
-		this.groupName = groupNameIn;
-		if (this.morph) {
-			this.morph.name = this.groupName;
-		}
-  }
-  
-  /**
-	 * Get the visibility of this Geometry.
-	 * 
-	 */
-	this.getVisibility = () => {
-		return this.morph.visible;
-	}
-	
-	/**
-	 * Set the visibility of this pointset.
-	 * @param {Boolean} flag - visibility to be set for this pointset.
-	 */
-	this.setVisibility = flag => {
-		this.morph.visible = flag;
-	}
-	
-	/**
-	 * Get the bounding box for the whole set of glyphs.
-	 * 
-	 * @return {Three.Box3};
-	 */
-	this.getBoundingBox = () => {
-		if (this.morph && this.morph.visible) {
-			//var boundingBox = new THREE.Box3().setFromObject(this.morph);
-      //return boundingBox;
-      let influences = this.morph.morphTargetInfluences;
-      let attributes = this.morph.geometry.morphAttributes;
-      if (influences && attributes && attributes.position) {
-        let min = new THREE.Vector3();
-        let max = new THREE.Vector3();
-        let found = false;
-        for (let i = 0; i < influences.length; i++) {
-          if (influences[i] > 0) {
-            found = true;
-            let box = new THREE.Box3().setFromArray(
-              attributes.position[i].array);
-            min.add(box.min.multiplyScalar(influences[i]));
-            max.add(box.max.multiplyScalar(influences[i]));
-          }
-        }
-        if (found) {
-          let boundingBox = new THREE.Box3(min, max);
-          return boundingBox;
-        }
-      }
-      var boundingBox = new THREE.Box3().setFromObject(this.morph);
-      return boundingBox;
-		}
-		return undefined;
-	}
-	
-	/**
-	 * Get the local time of this geometry, it returns a value between 
-	 * 0 and the duration.
-	 * 
-	 * @return {Number}
-	 */
-	this.getCurrentTime = () => {
-		if (this.clipAction) {
-			const ratio = this.clipAction.time / this.clipAction._clip.duration;
-			return this.duration * ratio;
-		} else {
-			return inbuildTime;
-		}
-	}
-
-	/**
-	 * Set the local time of this geometry.
-	 * 
-	 * @param {Number} time - Can be any value between 0 to duration.
-	 */
-	this.setMorphTime = time => {
-		if (this.clipAction) {
-			const ratio = time / this.duration;
-			const actualDuration = this.clipAction._clip.duration;
-			this.clipAction.time = ratio * actualDuration;
-			if (this.clipAction.time > actualDuration)
-				this.clipAction.time = actualDuration;
-			if (this.clipAction.time < 0.0)
-				this.clipAction.time = 0.0;
-			if (this.timeEnabled == 1)
-				this.mixer.update( 0.0 );
-		} else {
-			if (time > this.duration)
-				inbuildTime = this.duration;
-			else if (0 > time)
-				inbuildTime = 0;
-			else
-				inbuildTime = time;
-		}
-		if (this.morphColour == 1) {
-			if (typeof this.geometry !== "undefined") {
-				if (this.morph.material.vertexColors == THREE.VertexColors)
-				{
-					morphColorsToVertexColors(this.geometry, this.morph, this.clipAction)
-				}
-				this.geometry.colorsNeedUpdate = true;
-			}
-		}
-	}
-
-	/**
-	 * Check if the geometry is time varying.
-	 * 
-	 * @return {Boolean}
-	 */
-	this.isTimeVarying = () => {
-		if (this.timeEnabled || this.morphColour)
-			return true;
-		return false;
-	}
-
-		/**
-	 * Clear this geometry and free the memory.
-	 */
-	this.dispose = () => {
-		if (this.morph && this.morph.geometry)
-		  this.morph.geometry.dispose();
-		if (this.morph && this.morph.material)
-		  this.morph.material.dispose();
-		this.geometry = undefined;
-		this.mixer = undefined;
-		this.morph = undefined;
-		this.clipAction = undefined;
-		this.groupName = undefined;
-	}
-
-	//Update the geometry and colours depending on the morph.
-	this.render = (delta, playAnimation) => {
-		if (playAnimation == true) 
-		{
-			if ((this.clipAction) && (this.timeEnabled == 1)) {
-				this.mixer.update( delta );
-			}
-			else {
-				let targetTime = inbuildTime + delta;
-				if (targetTime > this.duration)
-					targetTime = targetTime - this.duration;
-				inbuildTime = targetTime;
-			}
-			if (this.morphColour == 1) {
-				if (typeof this.geometry !== "undefined") {
-					
-					if (this.morph.material.vertexColors == THREE.VertexColors)
-					{
-						let clipAction = undefined;
-						if (this.clipAction && (this.timeEnabled == 1))
-							clipAction = this.clipAction;
-						morphColorsToVertexColors(this.geometry, this.morph, clipAction);
-						this.geometry.colorsNeedUpdate = true;
-					}
-					
-				}
-			}	
-		}
-	}
-
 }
+
+Lines.prototype = Object.create((require('./primitives/zincObject').ZincObject).prototype);
+exports.Lines = Lines;
